requirements.txt
streamlit==1.45.1
pandas==2.3.0
numpy==2.3.0
Pillow==11.2.1
streamlit-calendar==1.3.1
duckdb==1.3.0